% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratTools-package.R
\docType{package}
\name{chevreulSeurat-package}
\alias{chevreulSeurat}
\alias{chevreulSeurat-package}
\title{chevreulSeurat: Tools for Managing Seurat Objects}
\description{
Tools for using seurat with single cell projects
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/whtns/chevreulSeurat}
  \item \url{https://whtns.github.io/chevreulSeurat/}
  \item Report bugs at \url{https://github.com/cobriniklab/chevreulSeurat/issues}
}

}
\author{
\strong{Maintainer}: Kevin Stachelek \email{kevin.stachelek@gmail.com} (\href{https://orcid.org/0000-0003-2085-695X}{ORCID})

Authors:
\itemize{
  \item Bhavana Bhat \email{bhavanabhat29@gmail.com}
}

Other contributors:
\itemize{
  \item Roman Hillje (Author of included Cerebro fragments) [contributor]
}

}
\keyword{internal}
