% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cols}
\alias{combine_cols}
\title{Reformat Seurat Object Metadata}
\usage{
combine_cols(seu, cols, new_col)
}
\arguments{
\item{seu}{A Seurat object}

\item{cols}{Columns}

\item{new_col}{New columns}
}
\description{
Reformat Seurat Object Metadata by Coalesced Columns
}
