% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{filter_merged_seu}
\alias{filter_merged_seu}
\title{Filter a Single Seurat Object}
\usage{
filter_merged_seu(seu, filter_var, filter_val, .drop = .drop)
}
\arguments{
\item{seu}{A seurat object}

\item{filter_var}{Variables to filter}

\item{filter_val}{Values to filter}

\item{.drop}{}
}
\description{
Filters a seurat object, retaining all the desired values of a variable
}
