% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_rows_to_top}
\alias{filter_rows_to_top}
\title{Filter Rows to Top}
\usage{
filter_rows_to_top(df, column, values)
}
\arguments{
\item{df}{A data frame}

\item{column}{Column containing the values of choice}

\item{values}{Values to be filtered to the top}
}
\description{
Filter rows that match the desired value to the top.
}
