% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_new_metadata}
\alias{format_new_metadata}
\title{Add Seurat Object Metadata}
\usage{
format_new_metadata(seu, datapath)
}
\arguments{
\item{seu}{A seurat object}

\item{datapath}{Path to file containing metadata}
}
\description{
Adds data to the seurat object to produce a seurat object with metadata added.
}
