% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ifnb1k}
\alias{ifnb1k}
\title{Interferon Beta-Stimulated and Control PBMCs (1k Cell Subset)}
\format{
A Seurat object with approximately 1000 cells and 14053 features.
The object contains metadata for stimulation condition ('stim') and cell
type annotations ('seurat_annotations').
}
\source{
\url{https://www.nature.com/articles/nbt.4042}
}
\usage{
ifnb1k
}
\description{
A subset of the dataset from Kang et al. (2018), containing peripheral blood
mononuclear cells (PBMCs) from 8 lupus patients, split into control and
interferon-beta (IFN-B) stimulated conditions. This \code{ifnb1k} object is a
randomly sampled subset of approximately 1000 cells used for demonstration
purposes in the seuratTools package.
}
\keyword{datasets}
