% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_seuFLViz_clean_names}
\alias{make_seuFLViz_clean_names}
\title{Clean Vector of seuFLViz Names}
\usage{
make_seuFLViz_clean_names(myvec)
}
\arguments{
\item{myvec}{A vector of seurat object names}
}
\description{
Cleans names of seurat objects provided in a vector form
}
