% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc1k}
\alias{pbmc1k}
\title{A random subset of 1000 cells from the PBMC 3K dataset}
\format{
A Seurat object containing 1000 cells.
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/pbmc3k}
}
\usage{
pbmc1k
}
\description{
A Seurat object containing 1000 randomly sampled cells from the pbmc3k dataset,
which is publicly available from 10X Genomics. This dataset is provided
as an example for the seuratTools package.
}
\keyword{datasets}
