% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_seurat.R
\name{process_seurat}
\alias{process_seurat}
\title{Basic processing of scRNA-Seq data in Seurat objects}
\usage{
process_seurat(
  obj,
  num_most_variable_features = 2000,
  logfile,
  outdir,
  vars_to_color = c("seurat_clusters"),
  image_save_mode = T,
  table_save_mode = T
)
}
\arguments{
\item{obj}{A seurat object}

\item{num_most_variable_features}{(Optional) input to FindVariableFeatures}

\item{logfile}{File to write console output to}

\item{outdir}{Output directory}

\item{vars_to_color}{Columns of the Seurat metadata to color the UMAP by}

\item{image_save_mode}{(Optional) Boolean indicating whether to save images}

\item{table_save_mode}{(Optional) Boolean indicating whether to save tables}
}
\value{
A processed seurat object
}
\description{
This function normalizes, scales data, and creates QC plots. Run this downstream of UMI and gene filtering + doublet detection
}
