% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seu_calcn}
\alias{seu_calcn}
\title{Calculate Read Count Metrics for a Seurat object}
\usage{
seu_calcn(seu, assay = "gene", slot = "counts")
}
\arguments{
\item{seu}{A seurat object}

\item{assay}{Assay to use, Default = "gene"}

\item{slot}{}
}
\description{
Recalculate counts/features per cell for a seurat object
}
