% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_integrate.R
\name{seurat_v5_integrate}
\alias{seurat_v5_integrate}
\title{Batch Correct Multiple Seurat Objects}
\usage{
seurat_v5_integrate(seu, method = "cca", organism = "human", ...)
}
\arguments{
\item{method}{Default "cca"}

\item{...}{}

\item{seu_list}{List of two or more seurat objects}
}
\description{
Batch Correct Multiple Seurat Objects
}
