% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_management.R
\name{subset_by_meta}
\alias{subset_by_meta}
\title{Subset by new metadata}
\usage{
subset_by_meta(meta_path, seu)
}
\arguments{
\item{meta_path}{Path to new metadata}

\item{seu}{A seurat obejct}
}
\description{
Subset the seurat object using new metadata
}
