% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_project_db}
\alias{create_project_db}
\title{Create a database of seuFLViz projects}
\usage{
create_project_db(
  cache_location = "~/.cache/seuFLViz",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{cache_location}{Path to cache "~/.cache/seuFLViz"}

\item{sqlite_db}{Database to be created}

\item{verbose}{}
}
\description{
Create a database containing seuFLViz projects
}
