% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_project_db}
\alias{read_project_db}
\title{Read a database of seuFLViz projects}
\usage{
read_project_db(
  projects_dir = NULL,
  cache_location = "~/.cache/seuFLViz",
  sqlite_db = "single-cell-projects.db",
  verbose = TRUE
)
}
\arguments{
\item{projects_dir}{Project directory containing seuFLViz projects}

\item{cache_location}{Path to cache "~/.cache/seuFLViz"}

\item{sqlite_db}{sqlite db}

\item{verbose}{}
}
\description{
Reads database of seuFLViz projects to a data frame
}
