% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_seuratTools_object}
\alias{update_seuratTools_object}
\title{Update a seuratTools Object}
\usage{
update_seuratTools_object(
  seu_path,
  feature,
  resolution = seq(0.2, 2, by = 0.2),
  return_seu = TRUE,
  ...
)
}
\arguments{
\item{seu_path}{Path to a seurat object}

\item{...}{}
}
\description{
Update a seuratTools Object
}
